requirements.txt
asgiref==3.4.1
Django==3.2.9
django-jazzmin==2.4.8
django-ninja==0.16.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
idna==3.3
Pillow==8.4.0
pyasn1==0.4.8
pydantic==1.8.2
python-jose==3.3.0
pytz==2021.3
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
